api made for theme to pick up (read) button settings and show buttons accordingly
