+++ abdullah-app/shopify.app.toml
@@ -3,7 +3,7 @@
 client_id = "fbf60f6302b3e92e01e42fa7dc123f70"
 name = "abdullah-app"
 handle = "abdullah-app-1"
-application_url = "https://7c28-122-129-85-58.ngrok-free.app"
+application_url = "https://7d13-122-129-85-58.ngrok-free.app"
 embedded = true
 
 [build]
@@ -17,9 +17,9 @@ scopes = "write_products"
 
 [auth]
 redirect_urls = [
-  "https://7c28-122-129-85-58.ngrok-free.app/auth/callback",
-  "https://7c28-122-129-85-58.ngrok-free.app/auth/shopify/callback",
-  "https://7c28-122-129-85-58.ngrok-free.app/api/auth/callback"
+  "https://7d13-122-129-85-58.ngrok-free.app/auth/callback",
+  "https://7d13-122-129-85-58.ngrok-free.app/auth/shopify/callback",
+  "https://7d13-122-129-85-58.ngrok-free.app/api/auth/callback"
 ]
 
 [webhooks]
